Gnn (#226)
* no message (#225)

* gnn graphs on entire datasets

* preparing gcn for merge

* functional gcn for homogeneous graph only and without test phase

* node2vec test

* plot embs

* no message

* no message

* no message

* no message

* gcn refactor

* Added gs_layer class which is similar to gcn_layer class, but contains the GraphSAGE model layers

* added gat_layer class

* Update readme.md

---------

+++ src/cmn/team.py
@@ -152,7 +152,7 @@ class Team(object):
         candidate_vec_dim = len(c2i)
         location_vec_dim = len(l2i)
         data = lil_matrix((len(teams), 1 + skill_vec_dim + candidate_vec_dim + location_vec_dim), dtype='u1')
-        data_ = np.zeros((bucket_size, 1 + skill_vec_dim + candidate_vec_dim + location_vec_dim))
+        data_ = np.zeros((bucket_size, 1 + skill_vec_dim + candidate_vec_dim + location_vec_dim), dtype='u1')
         j = -1
         st = time()
         for i, team in enumerate(teams):
